Refactored UI components
